###feat/Add feature extractors and test samples for SQLException
- Added KDoc comments with @since 1.3.0 and links to the sample file.
- Changed file naming to `sqlExceptionFeatureExtractors` and `SQLExceptionFeaturesSamples`
- Removed accidentally commited files
- Aligned comments and ordered fun
